requirements.txt
dominate==2.1.16
Flask==0.10.1
Flask-Bootstrap==3.3.5.6
Flask-Themes==0.1.3
Flask-WTF==0.12
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-instagram==1.3.2
simplejson==3.8.0
six==1.10.0
Werkzeug==0.10.4
wheel==0.24.0
WTForms==2.0.2
